homework1 fix remarks
