+++ homework1_1/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group = 'org.solodov'
+group = 'org.otus.solodov'
 version = '1.0-SNAPSHOT'
 
 repositories {
