+++ homework1_1/src/main/java/org/otus/solodov/TestResult.java
@@ -1,0 +1,7 @@
+package org.otus.solodov;
+
+public enum TestResult {
+    SUCCESS,
+    FAILURE,
+    DISABLED
+}
