+++ homework1_1/src/main/java/org/otus/solodov/annotation/Test.java
@@ -1,12 +1,12 @@
-package org.example.annotation;
+package org.otus.solodov.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.METHOD, ElementType.TYPE})
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Disabled {
-    String reason();
+public @interface Test {
+    int priority() default 5;
 }
