+++ homework1_1/src/main/java/org/otus/solodov/exception/IllegalAnnotationAttributeException.java
@@ -1,4 +1,4 @@
-package org.example.exception;
+package org.otus.solodov.exception;
 
 public class IllegalAnnotationAttributeException extends RuntimeException {
     public IllegalAnnotationAttributeException(String errorMessage) {
