+++ homework1_1/src/main/java/org/otus/solodov/exception/TestExecutionException.java
@@ -1,4 +1,4 @@
-package org.example.exception;
+package org.otus.solodov.exception;
 
 public class TestExecutionException extends RuntimeException {
     public TestExecutionException(String errorMessage) {
